Hookup the verifier
