+++ pkg/container/sigstore/sigstore.go
@@ -5,7 +5,11 @@ import (
 	"context"
 	"embed"
 	"errors"
+	"fmt"
+	"strings"
 
+	"github.com/google/go-containerregistry/pkg/name"
+	"github.com/google/go-containerregistry/pkg/v1/remote"
 	"github.com/sigstore/sigstore-go/pkg/root"
 	"github.com/sigstore/sigstore-go/pkg/verify"
 
@@ -16,10 +20,10 @@ import (
 var embeddedTufRoots embed.FS
 
 const (
-	// GitHubSigstoreTrustedRootRepo is the GitHub trusted root repository for sigstore
-	GitHubSigstoreTrustedRootRepo = "tuf-repo.github.com"
-	// SigstorePublicTrustedRootRepo is the public trusted root repository for sigstore
-	SigstorePublicTrustedRootRepo = "tuf-repo-cdn.sigstore.dev"
+	// TrustedRootSigstoreGitHub is the GitHub trusted root repository for sigstore
+	TrustedRootSigstoreGitHub = "tuf-repo.github.com"
+	// TrustedRootSigstorePublicGoodInstance is the public trusted root repository for sigstore
+	TrustedRootSigstorePublicGoodInstance = "tuf-repo-cdn.sigstore.dev"
 	// RootTUFPath is the path to the root.json file inside an embedded TUF repository
 	rootTUFPath = "root.json"
 )
@@ -31,7 +35,8 @@ type Sigstore struct {
 }
 
 // New creates a new Sigstore verifier
-func New(sigstoreTUFRepoURL string, authOpts ...AuthMethod) (*Sigstore, error) {
+func New(authOpts ...AuthMethod) (*Sigstore, error) {
+	sigstoreTUFRepoURL := TrustedRootSigstorePublicGoodInstance
 	// Get the sigstore options for the TUF client and the verifier
 	tufOpts, opts, err := getSigstoreOptions(sigstoreTUFRepoURL)
 	if err != nil {
@@ -56,19 +61,51 @@ func New(sigstoreTUFRepoURL string, authOpts ...AuthMethod) (*Sigstore, error) {
 	}, nil
 }
 
+func parseImageRef(imageRef string) (string, string, string, error) {
+	// Parse the image reference
+	ref, err := name.ParseReference(imageRef)
+	if err != nil {
+		return "", "", "", fmt.Errorf("error parsing image reference: %w", err)
+	}
+	switch r := ref.(type) {
+	case name.Tag:
+		// 0. Get the auth options
+		opts := []remote.Option{}
+
+		// 2. Get the image descriptor
+		desc, err := remote.Get(ref, opts...)
+		if err != nil {
+			return "", "", "", fmt.Errorf("error getting image descriptor: %w", err)
+		}
+
+		// 3. Get the digest
+		digest := ref.Context().Digest(desc.Digest.String())
+		owner := strings.Split(r.Context().Name(), "/")[0]
+		artifact := strings.Split(r.Context().Name(), "/")[1]
+		checksum := digest.Identifier()
+		return owner, artifact, checksum, nil
+	case name.Digest:
+		// strip the reference into 2 parts - the registry and everything else
+		parts := strings.SplitN(r.Context().Name(), "/", 2)
+		if len(parts) != 2 {
+			return "", "", "", fmt.Errorf("error: invalid image reference")
+		}
+		owner := parts[0]
+		artifact := strings.Split(strings.Join(parts[1:], "/"), "@")[0]
+		checksum := ref.Identifier()
+		return owner, artifact, checksum, nil
+	default:
+		return "", "", "", fmt.Errorf("error: unknown image reference type")
+	}
+}
+
 // Verify verifies a container artifact using sigstore
 func (s *Sigstore) Verify(
 	ctx context.Context,
-	sev *verify.SignedEntityVerifier,
-	owner, artifact, checksumref string,
+	imageRef string,
 ) ([]Result, error) {
-	// Sanitize the input
-	sanitizeInput(&owner)
-
-	cauth := newContainerAuth(s.authOpts...)
-
 	// Construct the bundle(s) - OCI image or GitHub's attestation endpoint
-	bundles, err := getSigstoreBundles(ctx, owner, artifact, checksumref, cauth)
+	bundles, err := getSigstoreBundles(ctx, imageRef)
 	if err != nil && !errors.Is(err, ErrProvenanceNotFoundOrIncomplete) {
 		// We got some other unexpected error prior to querying for the signature/attestation
 		return nil, err
@@ -86,5 +123,5 @@ func (s *Sigstore) Verify(
 	}
 
 	// Construct the verification result for each bundle we managed to generate.
-	return getVerifiedResults(ctx, sev, bundles), nil
+	return getVerifiedResults(ctx, s.verifier, bundles), nil
 }
